climagen/server/requirements.txt
openai
fastapi
uvicorn
python-dotenv
firebase-admin==6.5.0
pandas==2.2.1
pyarrow==15.0.0
langchain==0.0.340
langchain-core==0.1.23
chromadb==0.4.18
tiktoken==0.5.1